% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agpop.R
\docType{data}
\name{agpop}
\alias{agpop}
\title{agpop data}
\format{
This data frame contains the following columns:

\describe{
\item{county:}{county name (character variable)}

\item{state:}{state abbreviation (character variable)}

\item{acres92:}{number of acres devoted to farms, 1992}

\item{acres87:}{number of acres devoted to farms, 1987}

\item{acres82:}{number of acres devoted to farms, 1982}

\item{farms92:}{number of farms, 1992}

\item{farms87:}{number of farms, 1987}

\item{farms82:}{number of farms, 1982}

\item{largef92:}{number of farms with 1,000 acres or more, 1992}

\item{largef87:}{number of farms with 1,000 acres or more, 1987}

\item{largef82:}{number of farms with 1,000 acres or more, 1982}

\item{smallf92:}{number of farms with 9 acres or fewer, 1992}

\item{smallf87:}{number of farms with 9 acres or fewer, 1987}

\item{smallf82:}{number of farms with 9 acres or fewer, 1982}

\item{region:}{S = south; W = west; NC = north central; NE = northeast}
}
}
\usage{
data(agpop)
}
\description{
Data from the 1992 U.S. Census of Agriculture.
}
\references{
Lohr (2021), Sampling: Design and Analysis, 3rd Edition. Boca Raton, FL: CRC Press.

Lu and Lohr (2021), R Companion for \emph{Sampling: Design and Analysis, 3rd Edition}, 1st Edition. Boca Raton, FL: CRC Press.
}
\keyword{datasets}
