% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agpps.R
\docType{data}
\name{agpps}
\alias{agpps}
\title{agpps data}
\format{
This data frame contains the following columns:

\describe{
\item{county:}{county name (character variable)}

\item{state:}{state abbreviation (character variable)}

\item{acres92:}{number of acres devoted to farms, 1992}

\item{acres87:}{number of acres devoted to farms, 1987}


\item{acres82:}{number of acres devoted to farms, 1982}

\item{farms92:}{number of farms, 1992}

\item{farms87:}{number of farms, 1987}

\item{farms82:}{number of farms, 1982}

\item{largef92:}{number of farms with 1,000 acres or more, 1992}

\item{largef87:}{number of farms with 1,000 acres or more, 1987}

\item{largef82:}{number of farms with 1,000 acres or more, 1982}

\item{smallf92:}{number of farms with 9 acres or fewer, 1992}

\item{smallf87:}{number of farms with 9 acres or fewer, 1987}


\item{smallf82:}{number of farms with 9 acres or fewer, 1982}

\item{region:}{S = south; W = west; NC = north central; NE = northeast}

\item{sizemeas:}{size measure used to select the pps sample}
 
\item{SelectionProb:}{inclusion probability for county}
 
\item{SamplingWeight:}{sampling weight for county}
 
\item{Unit:}{unit number for indexing joint inclusion probabilities}
 
\item{JtProb_1:}{columns of joint inclusion probabilities}

\item{JtProb_2:}{columns of joint inclusion probabilities}

\item{JtProb_3:}{columns of joint inclusion probabilities}

\item{JtProb_4:}{columns of joint inclusion probabilities}

\item{JtProb_5:}{columns of joint inclusion probabilities}

\item{JtProb_6:}{columns of joint inclusion probabilities}

\item{JtProb_7:}{columns of joint inclusion probabilities}

\item{JtProb_8:}{columns of joint inclusion probabilities}

\item{JtProb_9:}{columns of joint inclusion probabilities}

\item{JtProb_10:}{columns of joint inclusion probabilities}

\item{JtProb_11:}{columns of joint inclusion probabilities}

\item{JtProb_12:}{columns of joint inclusion probabilities}

\item{JtProb_13:}{columns of joint inclusion probabilities}

\item{JtProb_14:}{columns of joint inclusion probabilities}

\item{JtProb_15:}{columns of joint inclusion probabilities}
}
}
\usage{
data(agpps)
}
\description{
Data from a without-replacement probability-proportional-to-size sample from \emph{agpop} data.
}
\references{
Lohr (2021), Sampling: Design and Analysis, 3rd Edition. Boca Raton, FL: CRC Press.

Lu and Lohr (2021), R Companion for \emph{Sampling: Design and Analysis, 3rd Edition}, 1st Edition. Boca Raton, FL: CRC Press.
}
\keyword{datasets}
