% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agsrs.R
\docType{data}
\name{agsrs}
\alias{agsrs}
\title{agsrs data}
\format{
Variables are the same as in \emph{agpop} data.
}
\usage{
data(agsrs)
}
\description{
Data from an SRS of size 300 from the 1992 U.S. Census of Agriculture \emph{agpop} data.
}
\references{
Lohr (2021), Sampling: Design and Analysis, 3rd Edition. Boca Raton, FL: CRC Press.

Lu and Lohr (2021), R Companion for \emph{Sampling: Design and Analysis, 3rd Edition}, 1st Edition. Boca Raton, FL: CRC Press.
}
\keyword{datasets}
