% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agstrat.R
\docType{data}
\name{agstrat}
\alias{agstrat}
\title{agstrat data}
\format{
This data frame contains the following columns:

county:   county name (character variable)

state: state abbreviation (character variable)

acres92:number of acres devoted to farms, 1992

acres87:number of acres devoted to farms, 1987

acres82:number of acres devoted to farms, 1982

farms92:number of farms, 1992

farms87:number of farms, 1987

farms82:number of farms, 1982

largef92:number of farms with 1,000 acres or more, 1992

largef87:number of farms with 1,000 acres or more, 1987

largef82:number of farms with 1,000 acres or more, 1982

smallf92:number of farms with 9 acres or fewer, 1992

smallf87:number of farms with 9 acres or fewer, 1987

smallf82:number of farms with 9 acres or fewer, 1982

region:S = south; W = west; NC = north central; NE = northeast

rn: random numbers used to select sample in each stratum  
 
weight: sampling weight for each county in sample
}
\usage{
data(agstrat)
}
\description{
Data from a stratified random sample of size 300 from the 1992 U.S. Census of Agriculture
}
\references{
Lohr (2021). Sampling: Design and Analysis, 3rd Edition.
}
\keyword{datasets}
