% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algebra.R
\docType{data}
\name{algebra}
\alias{algebra}
\title{algebra data}
\format{
This data frame contains the following columns:

class: Class number

Mi: Number of students (M i ) in class

score: Score of student on test
}
\usage{
data(algebra)
}
\description{
Hypothetical data for an SRS of 12 algebra classes in a city, from a population of 187 classes.
}
\references{
Lohr (2021). Sampling: Design and Analysis, 3rd Edition.
}
\keyword{datasets}
