% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anthrop.R
\docType{data}
\name{anthrop}
\alias{anthrop}
\title{anthrop data}
\format{
This data frame contains the following columns:

\describe{
\item{finger:}{length of left middle finger (cm)}

\item{height:}{height (inches)}
}
}
\usage{
data(anthrop)
}
\description{
Finger length and height for 3,000 criminals. This
data set contains information for the entire population.
}
\references{
Macdonell, W. R. (1901). On criminal anthropometry and the identification of criminals.
\emph{Biometrika 1}, 177–227.

Lohr (2021), Sampling: Design and Analysis, 3rd Edition. Boca Raton, FL: CRC Press.

Lu and Lohr (2021), R Companion for \emph{Sampling: Design and Analysis, 3rd Edition}, 1st Edition. Boca Raton, FL: CRC Press.
}
\keyword{datasets}
