% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anthsrs.R
\docType{data}
\name{anthsrs}
\alias{anthsrs}
\title{anthsrs data}
\format{
This data frame contains the following columns:

\describe{
\item{finger:}{length of left middle finger (cm)}
  
\item{height:}{height (inches)}
 
\item{wt:}{sampling weight}
}
}
\usage{
data(anthsrs)
}
\description{
Length of left middle finger and height for an SRS of size 200 from \emph{anthrop} data.
}
\references{
Lohr (2021), Sampling: Design and Analysis, 3rd Edition. Boca Raton, FL: CRC Press.

Lu and Lohr (2021), R Companion for \emph{Sampling: Design and Analysis, 3rd Edition}, 1st Edition. Boca Raton, FL: CRC Press.
}
\keyword{datasets}
