% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anthuneq.R
\docType{data}
\name{anthuneq}
\alias{anthuneq}
\title{anthuneq data}
\format{
This data frame contains the following columns:
\describe{
\item{finger:}{length of left middle finger (cm)}

\item{height:}{height (inches)}
 
\item{wt:}{sampling weight}
}
}
\usage{
data(anthuneq)
}
\description{
Finger length and height for a with replacement unequal probability sample of size 200 
from data \emph{anthrop}. The probability of selection, \eqn{\psi_i}, was proportional to 24 for
y < 65 , 12 for y = 65, 2 for y = 66 or 67, and 1 for y > 67.
}
\references{
Lohr (2021), Sampling: Design and Analysis, 3rd Edition. Boca Raton, FL: CRC Press.

Lu and Lohr (2021), R Companion for \emph{Sampling: Design and Analysis, 3rd Edition}, 1st Edition. Boca Raton, FL: CRC Press.
}
\keyword{datasets}
