% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditresult.R
\docType{data}
\name{auditresult}
\alias{auditresult}
\title{auditresult data}
\format{
This data frame contains the following columns:

\describe{
\item{account:}{audit unit}

\item{bookvalue:}{book value of account}

\item{psi:}{probability of selection}


\item{auditvalue:}{audit value of account}
}
}
\usage{
data(auditresult)
}
\description{
Audit data used in Chapter 6 of SDA.
}
\references{
Lohr (2021), Sampling: Design and Analysis, 3rd Edition. Boca Raton, FL: CRC Press.

Lu and Lohr (2021), R Companion for \emph{Sampling: Design and Analysis, 3rd Edition}, 1st Edition. Boca Raton, FL: CRC Press.
}
\keyword{datasets}
