% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditselect.R
\docType{data}
\name{auditselect}
\alias{auditselect}
\title{auditselect data}
\format{
This data frame contains the following columns:

\describe{
\item{account:}{audit unit}

\item{bookval:}{book value of account}
 
\item{cumbv:}{cumulative book value}

\item{rn1:}{random number 1 selecting account}

 \item{rn2:}{random number 2 selecting account}
 
  \item{rn3:}{random number 3 selecting account}
  }
}
\usage{
data(auditselect)
}
\description{
Selection of accounts for \emph{audit} data used in Chapter 6 of SDA.
}
\references{
Lohr (2021), Sampling: Design and Analysis, 3rd Edition. Boca Raton, FL: CRC Press.

Lu and Lohr (2021), R Companion for \emph{Sampling: Design and Analysis, 3rd Edition}, 1st Edition. Boca Raton, FL: CRC Press.
}
\keyword{datasets}
