% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azcounties.R
\docType{data}
\name{azcounties}
\alias{azcounties}
\title{azcounties data}
\format{
This data frame contains the following columns:

\describe{

\item{name:}{county name (character variable, length 15)}

\item{number:}{county number}

\item{population:}{population estimate for county}

\item{housing:}{housing unit estimate for county}

\item{ownerocc:}{number of owner-occupied housing units for county}
}
}
\usage{
data(azcounties)
}
\description{
Population and housing unit estimates for Arizona counties, excluding
Maricopa and Pima counties, from the American Community Survey 2018 5-year estimates.
}
\references{
Source: https://data.census.gov/, accessed November 27, 2020.

Lohr (2021), Sampling: Design and Analysis, 3rd Edition. Boca Raton, FL: CRC Press.

Lu and Lohr (2021), R Companion for \emph{Sampling: Design and Analysis, 3rd Edition}, 1st Edition. Boca Raton, FL: CRC Press.
}
\keyword{datasets}
