% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census1920.R
\docType{data}
\name{census1920}
\alias{census1920}
\title{census1920 data}
\format{
This data frame contains the following columns:

\describe{
\item{state:}{state name}

\item{population:}{state population in 1920 census}
}
}
\usage{
data(census1920)
}
\description{
Population sizes for each state, from the 1920 U.S. census. The data
set contains only the 48 states and excludes Washington D.C., Puerto Rico, and U.S.
territories (these areas were not allowed to have voting representatives in Congress).
}
\references{
Source: U.S. Bureau of the Census (1921).

Lohr (2021), Sampling: Design and Analysis, 3rd Edition. Boca Raton, FL: CRC Press.

Lu and Lohr (2021), R Companion for \emph{Sampling: Design and Analysis, 3rd Edition}, 1st Edition. Boca Raton, FL: CRC Press.
}
\keyword{datasets}
