% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census2010.R
\docType{data}
\name{census2010}
\alias{census2010}
\title{census2010 data}
\format{
This data frame contains the following columns:

\describe{
\item{state:}{state name}

\item{population:}{state population in 2010 census}
}
}
\usage{
data(census2010)
}
\description{
Population sizes for each state, from the 2010 U.S. census. The data set contains only the 50 states and excludes the
areas that, as of 2020, are not allowed to have voting representatives in Congress: 
Washington D.C., Puerto Rico, and U.S. territories.
}
\references{
Source: U.S. Census Bureau (2019).

Lohr (2021), Sampling: Design and Analysis, 3rd Edition. Boca Raton, FL: CRC Press.

Lu and Lohr (2021), R Companion for \emph{Sampling: Design and Analysis, 3rd Edition}, 1st Edition. Boca Raton, FL: CRC Press.
}
\keyword{datasets}
