% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cherry.R
\docType{data}
\name{cherry}
\alias{cherry}
\title{cherry data}
\format{
This data frame contains the following columns:

\describe{
\item{diameter:}{diameter of tree (inches)}

\item{height:}{height of tree (feet)}

\item{volume:}{timber volume of tree (cubic feet)}
}
}
\usage{
data(cherry)
}
\description{
Data for a sample of 31 cherry trees.
}
\references{
Hand, D. J., F. Daly, A. D. Lunn, K. J. McConway, and E. Ostrowski (1994). A Handbook
of Small Data Sets. London: Chapman and Hall.

Lohr (2021), Sampling: Design and Analysis, 3rd Edition. Boca Raton, FL: CRC Press.

Lu and Lohr (2021), R Companion for \emph{Sampling: Design and Analysis, 3rd Edition}, 1st Edition. Boca Raton, FL: CRC Press.
}
\keyword{datasets}
