% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{data}
\name{classes}
\alias{classes}
\title{classes data}
\format{
This data frame contains the following columns:

class: Class ID number

class_size: Number of students in class
}
\usage{
data(classes)
}
\description{
Population sizes for 15 classes, used in Chapter 6 of SDA to demonstrate
unequal-probability sampling.
}
\references{
Lohr (2021). Sampling: Design and Analysis, 3rd Edition.
}
\keyword{datasets}
