% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classpps.R
\docType{data}
\name{classpps}
\alias{classpps}
\title{classpps data}
\format{
This data frame contains the following columns:

\describe{
\item{class:}{class ID number}

\item{class_size:}{number of students in class}

\item{finalweight:}{sampling weight for student}

\item{hours:}{number of hours spent studying statistics}
}
}
\usage{
data(classpps)
}
\description{
Two-stage unequal-probability sample without replacement from the population of classes in \emph{classes} data.
}
\references{
Lohr (2021), Sampling: Design and Analysis, 3rd Edition. Boca Raton, FL: CRC Press.

Lu and Lohr (2021), R Companion for \emph{Sampling: Design and Analysis, 3rd Edition}, 1st Edition. Boca Raton, FL: CRC Press.
}
\keyword{datasets}
