% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classppsjp.R
\docType{data}
\name{classppsjp}
\alias{classppsjp}
\title{classppsjp data}
\format{
This data frame contains the following columns:

\describe{
\item{class:}{class ID number}

\item{class_size:}{number of students in class}

\item{SelectionProb:}{probability of being included in sample, \eqn{\pi_i}}

\item{SamplingWeight:}{sampling weight \eqn{w_i = 1/(\pi_i)}}

\item{JtProb_1:}{columns of joint inclusion probabilities, \eqn{\pi_{1k}}}

\item{JtProb_2:}{columns of joint inclusion probabilities, \eqn{\pi_{2k}}}

\item{JtProb_3:}{columns of joint inclusion probabilities, \eqn{\pi_{3k}}}

\item{JtProb_4:}{columns of joint inclusion probabilities, \eqn{\pi_{4k}}}

\item{JtProb_5:}{columns of joint inclusion probabilities, \eqn{\pi_{5k}}}
}
}
\usage{
data(classppsjp)
}
\description{
Joint inclusion probabilities for unequal probability sample without replacement from
 the population of classes in data \emph{classes}.
}
\references{
Lohr (2021), Sampling: Design and Analysis, 3rd Edition. Boca Raton, FL: CRC Press.

Lu and Lohr (2021), R Companion for \emph{Sampling: Design and Analysis, 3rd Edition}, 1st Edition.
Boca Raton, FL: CRC Press.
}
\keyword{datasets}
