% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golfsrs.R
\docType{data}
\name{golfsrs}
\alias{golfsrs}
\title{golfsrs data}
\format{
This data frame contains the following columns:

\describe{
\item{RN:}{random number used to select golf course for sample}

\item{state:}{state name}

\item{holes:}{number of holes}

\item{type:}{type of course: 

priv = private 

semi = semi-private

pub = public 

mili = military

resort = resort}
 
\item{yearblt:}{year course was built}
 
\item{wkday18:}{greens fee for 18 holes during week}
 
\item{wkday9:}{greens fee for 9 holes during week}
 
\item{wkend18:}{greens fee for 18 holes on weekend}
 
\item{wkend9:}{greens fee for 9 holes on weekend}
 
\item{backtee:}{back tee yardage}
 
\item{rating:}{course rating}
 
\item{par:}{par for course}
 
\item{cart18:}{golf cart rental fee for 18 holes}
 
\item{cart9:}{golf cart rental fee for 9 holes}
 
\item{caddy:}{are caddies available? (y or n)}
 
\item{pro:}{is a golf pro available? (y or n)}
}
}
\usage{
data(golfsrs)
}
\description{
A simple random sample of 120 golf courses, taken from the population on the
website ww2.golfcourse.com on August 5, 1998. Missing data are denoted by NA.
}
\references{
Lohr (2021), Sampling: Design and Analysis, 3rd Edition. Boca Raton, FL: CRC Press.

Lu and Lohr (2021), R Companion for \emph{Sampling: Design and Analysis, 3rd Edition}, 1st Edition. Boca Raton, FL: CRC Press.
}
\keyword{datasets}
