% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpa.R
\docType{data}
\name{gpa}
\alias{gpa}
\title{gpa data}
\format{
This data frame contains the following columns:

suite: Suite (psu) identifier

gpa: Grade point average of person in suite

wt: Sampling weight, = 20 for every observation
}
\usage{
data(gpa)
}
\description{
GPA data from Chapter 5 of SDA.
}
\references{
Lohr (2021). Sampling: Design and Analysis, 3rd Edition.
}
\keyword{datasets}
