% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htpop.R
\docType{data}
\name{htpop}
\alias{htpop}
\title{htpop data}
\format{
This data frame contains the following columns:

\describe{
\item{height:}{height of person, cm}

\item{gender:}{
M = male

F = female}
}
}
\usage{
data(htpop)
}
\description{
Height and gender of 2,000 persons in an artificial population.
}
\references{
Lohr (2021), Sampling: Design and Analysis, 3rd Edition. Boca Raton, FL: CRC Press.

Lu and Lohr (2021), R Companion for \emph{Sampling: Design and Analysis, 3rd Edition}, 1st Edition. Boca Raton, FL: CRC Press.
}
\keyword{datasets}
