% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htsrs.R
\docType{data}
\name{htsrs}
\alias{htsrs}
\title{htsrs data}
\format{
This data frame contains the following columns:

rn: random number used to select unit

height: height of person, cm

gender: M=male, F=female
}
\usage{
data(htsrs)
}
\description{
Height and gender for a SRS of 200 persons, taken from htpop data
}
\references{
Lohr (2021). Sampling: Design and Analysis, 3rd Edition.
}
\keyword{datasets}
