% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hunting.R
\docType{data}
\name{hunting}
\alias{hunting}
\title{hunting data}
\format{
This data frame contains the following columns:
\describe{
\item{region:}{region of state (East, Central, West)}

\item{gender:}{gender (female, male)}

\item{age:}{age group (16-24, 25-34, 35-44, 45-54, 55-64, 65+)}

\item{popsize:}{population size in poststratum from the 2000 U.S. census}

\item{sampsize:}{sample size in poststratum}
}
}
\usage{
data(hunting)
}
\description{
Population and sample sizes for the poststrata used for the Sunday hunting
survey.
}
\references{
Source: Virginia Polytechnic and State University/Responsive Management (2006).

Lohr (2021), Sampling: Design and Analysis, 3rd Edition. Boca Raton, FL: CRC Press.

Lu and Lohr (2021), R Companion for \emph{Sampling: Design and Analysis, 3rd Edition}, 1st Edition. Boca Raton, FL: CRC Press.
}
\keyword{datasets}
