% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\docType{data}
\name{impute}
\alias{impute}
\title{impute data}
\format{
This data frame contains the following columns:
\describe{
\item{person:}{identification number for person}

\item{age:}{age in years}

\item{gender:}{
M = male

 F = female}

\item{education:}{number of years of education}

\item{crime:}{
= 1 if victim of any crime

= 0 otherwise}

\item{violcrime:}{
= 1 if victim of violent crime

= 0 otherwise}
}
}
\usage{
data(impute)
}
\description{
Small artificial data set used to illustrate imputation methods. Missing values
are denoted by NA.
}
\references{
Lohr (2021), Sampling: Design and Analysis, 3rd Edition. Boca Raton, FL: CRC Press.

Lu and Lohr (2021), R Companion for \emph{Sampling: Design and Analysis, 3rd Edition}, 1st Edition. Boca Raton, FL: CRC Press.
}
\keyword{datasets}
