% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integerwt.R
\docType{data}
\name{integerwt}
\alias{integerwt}
\title{integerwt data}
\format{
This data frame contains the following columns:
\describe{
\item{stratum:}{stratum number}
 
 \item{y:}{\eqn{y} value of observation}
 }
}
\usage{
data(integerwt)
}
\description{
Artificial population of 2000 observations.
}
\references{
Lohr (2021), Sampling: Design and Analysis, 3rd Edition. Boca Raton, FL: CRC Press.

Lu and Lohr (2021), R Companion for \emph{Sampling: Design and Analysis, 3rd Edition}, 1st Edition. Boca Raton, FL: CRC Press.
}
\keyword{datasets}
