% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intellonline.R
\docType{data}
\name{intellonline}
\alias{intellonline}
\title{intellonline data}
\format{
This data frame contains the following columns:

\describe{
\item{int:}{response to question about agreement with the statement "I
am more intelligent than the average person"

 1 = Strongly Agree 
 
 2 = Mostly Agree 
 
 3 = Mostly Disagree
 
4 = Strongly Disagree 

5 = Don't Know or Not Sure}

\item{region:}{census region of respondent (character variable, length 10): 

Northeast

South
 
Midwest
   
West}

\item{sex:}{sex (character variable, length 8): 

Male

Female}

\item{race:}{race (character variable, length 18): 

White

African American

Asian American

Hispanic American

 Another origin}

\item{age:}{age, years}
\item{income:}{household income level (character variable, length 8):

 < $40k,
 
 $40-80k, 
 
  > $80k}
 
 \item{education:}{highest education level attained (character variable, length
12): 

No College

  Some College
  
    College Grad
    
      Grad School
      
      MISSING}
 
\item{postwt:}{relative weight, obtained by poststratifying to demographic
proportions in the 2010 U.S. Census. The weights are normed
so that they sum to 750.}
}
}
\usage{
data(intellonline)
}
\description{
Data from the online (Mechanical Turk) survey.
The data were downloaded from \href{https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0200103}{PLOS ONE} on February 8, 2020; the variables extracted from the full
data set are provided here for educational purposes only.
}
\references{
Heck, P. R., D. J. Simons, and C. F. Chabris (2018). 65\% of Americans believe they
are above average in intelligence: Results of two nationally representative surveys. \emph{PloS
One 13 (7)}, 1–11.

Lohr (2021), Sampling: Design and Analysis, 3rd Edition. Boca Raton, FL: CRC Press.

Lu and Lohr (2021), R Companion for \emph{Sampling: Design and Analysis, 3rd Edition}, 1st Edition. Boca Raton, FL: CRC Press.
}
\keyword{datasets}
