% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intelltel.R
\docType{data}
\name{intelltel}
\alias{intelltel}
\title{intelltel data}
\format{
The
variables are the same as in \emph{intellonline}.
}
\usage{
data(intelltel)
}
\description{
Data from the telephone survey studied by Heck et al. (2018). The
data were downloaded from \href{https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0200103}{here} 
and are provided for educational purposes only.
}
\references{
Heck, P. R., D. J. Simons, and C. F. Chabris (2018). 65\% of Americans believe they
are above average in intelligence: Results of two nationally representative surveys. \emph{PloS One 13 (7)}, 1–11.

Lohr (2021), Sampling: Design and Analysis, 3rd Edition. Boca Raton, FL: CRC Press.

Lu and Lohr (2021), R Companion for \emph{Sampling: Design and Analysis, 3rd Edition}, 1st Edition. Boca Raton, FL: CRC Press.
}
\keyword{datasets}
