% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otters.R
\docType{data}
\name{otters}
\alias{otters}
\title{otters data}
\format{
This data frame contains the following columns:

\describe{
\item{section:}{section of coastline}

\item{habitat:}{type of habitat (stratum)}

\item{holts:}{number of holts (dens)}
}
}
\usage{
data(otters)
}
\description{
Data on number of holts (dens) in Shetland, U.K., used in Kruuk et al. (1989).
Data courtesy of Hans Kruuk.
}
\references{
Kruuk et al. (1989) An estimate of numbers and habitat preferences of otters Lutra lutra in Shetland, UK. \emph{Biological
Conservation 49}, 241–254.

Lohr (2021), Sampling: Design and Analysis, 3rd Edition. Boca Raton, FL: CRC Press.

Lu and Lohr (2021), R Companion for \emph{Sampling: Design and Analysis, 3rd Edition}, 1st Edition. Boca Raton, FL: CRC Press.
}
\keyword{datasets}
