% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radon.R
\docType{data}
\name{radon}
\alias{radon}
\title{radon data}
\format{
This data frame contains the following columns:

\describe{
\item{countyname:}{county name}

\item{countynum:}{county number}

\item{sampsize:}{sample size in county}

\item{popsize:}{population size in county}

\item{radon:}{radon concentration (picocuries per liter)}
 
 }
}
\usage{
data(radon)
}
\description{
Radon readings for a stratified sample of 1003 homes in Minnesota. The data were downloaded in April 2008 from an earlier version
of the website now located at \cr
\href{https://www.stat.berkeley.edu/users/statlabs/labs.html}{https://www.stat.berkeley.edu/users/statlabs/labs.html}.
}
\references{
Lohr (2021), Sampling: Design and Analysis, 3rd Edition. Boca Raton, FL: CRC Press.

Lu and Lohr (2021), R Companion for \emph{Sampling: Design and Analysis, 3rd Edition}, 1st Edition. Boca Raton, FL: CRC Press.
}
\keyword{datasets}
