% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectlength.R
\docType{data}
\name{rectlength}
\alias{rectlength}
\title{rectlength data}
\format{
This data frame contains the following columns:

\describe{
\item{rectangle:}{rectangle number}

\item{length:}{rectangle length}
 }
}
\usage{
data(rectlength)
}
\description{
Lengths of rectangles.
}
\references{
Lohr (2021), Sampling: Design and Analysis, 3rd Edition. Boca Raton, FL: CRC Press.

Lu and Lohr (2021), R Companion for \emph{Sampling: Design and Analysis, 3rd Edition}, 1st Edition. Boca Raton, FL: CRC Press.
}
\keyword{datasets}
