% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnt.R
\docType{data}
\name{rnt}
\alias{rnt}
\title{rnt data}
\format{
This data frame contains the following columns:
\describe{
\item{col1:}{column of 5-digit random numbers}

\item{col2:}{column of 5-digit random numbers}

\item{col3:}{column of 5-digit random numbers}

\item{col4:}{column of 5-digit random numbers}

\item{col5:}{column of 5-digit random numbers}

\item{col6:}{column of 5-digit random numbers}
}
}
\usage{
data(rnt)
}
\description{
Page from a random number table.
}
\references{
Lohr (2021), Sampling: Design and Analysis, 3rd Edition. Boca Raton, FL: CRC Press.

Lu and Lohr (2021), R Companion for \emph{Sampling: Design and Analysis, 3rd Edition}, 1st Edition. Boca Raton, FL: CRC Press.
}
\keyword{datasets}
