% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/santacruz.R
\docType{data}
\name{santacruz}
\alias{santacruz}
\title{santacruz data}
\format{
This data frame contains the following columns:
\describe{
\item{tree:}{tree number}

\item{seed92:}{number of seedlings in 1992}

\item{seed94:}{number of seedlings in 1994}
}
}
\usage{
data(santacruz)
}
\description{
The number of seedlings in the sampled psus on Santa Cruz Island, California, in 1992 and 1994.
}
\references{
Peart, D. (1994). Impacts of Feral Pig Activity on Vegetation Patterns Associated with
Quercus agrifolia on Santa Cruz Island, California. \emph{Ph.D. dissertation}. Tempe, AZ:
Arizona State University.

Lohr (2021), Sampling: Design and Analysis, 3rd Edition. Boca Raton, FL: CRC Press.

Lu and Lohr (2021), R Companion for \emph{Sampling: Design and Analysis, 3rd Edition}, 1st Edition. Boca Raton, FL: CRC Press.
}
\keyword{datasets}
