% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seals.R
\docType{data}
\name{seals}
\alias{seals}
\title{seals data}
\format{
This data frame contains the following columns:
\describe{
\item{zone:}{zone number for sampled area}

\item{holes:}{number of breathing holes Imjak found in area}
}
}
\usage{
data(seals)
}
\description{
Data on number of breathing holes found in sampled areas of Svalbard fjords,
reconstructed from summary statistics given in Lydersen and Ryg (1991).
}
\references{
Lydersen, C. and M. Ryg (1991). Evaluating breeding habitat and populations of ringed
seals Phoca hispida in Svalbard fjords. \emph{Polar Record 27}, 223–228.

Lohr (2021), Sampling: Design and Analysis, 3rd Edition. Boca Raton, FL: CRC Press.

Lu and Lohr (2021), R Companion for \emph{Sampling: Design and Analysis, 3rd Edition}, 1st Edition. Boca Raton, FL: CRC Press.
}
\keyword{datasets}
