% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapespop.R
\docType{data}
\name{shapespop}
\alias{shapespop}
\title{shapespop data}
\format{
This data frame contains the following columns:
\describe{
\item{ID:}{identification number for object}

\item{shape:}{shape of object (square or circle)}

\item{color:}{color of object (gray or black)}

\item{area:}{area of object (\eqn{cm^2})}

\item{conv:}{= 1 if object can be reached through convenience sample

 = 0 otherwise}
}
}
\usage{
data(shapespop)
}
\description{
Population of black and gray squares and circles.
}
\references{
Lohr (2021), Sampling: Design and Analysis, 3rd Edition. Boca Raton, FL: CRC Press.

Lu and Lohr (2021), R Companion for \emph{Sampling: Design and Analysis, 3rd Edition}, 1st Edition. Boca Raton, FL: CRC Press.
}
\keyword{datasets}
