% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanish.R
\docType{data}
\name{spanish}
\alias{spanish}
\title{spanish data}
\format{
This data frame contains the following columns:

\describe{
\item{class:}{class number}

\item{score:}{score on vocabulary test (out of 100)}

\item{trip:}{= 1 if plan a trip to a Spanish-speaking country

 = 0 otherwise}
}
}
\usage{
data(spanish)
}
\description{
Fictional cluster sample of introductory Spanish students.
}
\references{
Lohr (2021), Sampling: Design and Analysis, 3rd Edition. Boca Raton, FL: CRC Press.

Lu and Lohr (2021), R Companion for \emph{Sampling: Design and Analysis, 3rd Edition}, 1st Edition. Boca Raton, FL: CRC Press.
}
\keyword{datasets}
