% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srs30.R
\docType{data}
\name{srs30}
\alias{srs30}
\title{srs30 data}
\format{
This data frame contains the following columns:

\describe{
\item{y:}{value of observation}
}
}
\usage{
data(srs30)
}
\description{
An SRS of size 30 taken from an artificial population of size 100.
}
\references{
Lohr (2021), Sampling: Design and Analysis, 3rd Edition. Boca Raton, FL: CRC Press.

Lu and Lohr (2021), R Companion for \emph{Sampling: Design and Analysis, 3rd Edition}, 1st Edition. Boca Raton, FL: CRC Press.
}
\keyword{datasets}
