% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssc.R
\docType{data}
\name{ssc}
\alias{ssc}
\title{ssc data}
\format{
This data frame contains the following columns:
\describe{
\item{gender:}{m = male

  f = female}

\item{occupation:}{a = academic
  
 i = industry
 
 n = not determined}
 
 \item{ASA:}{= 1 if person is member of American Statistical Association
 
  = 0 otherwise}
 }
}
\usage{
data(ssc)
}
\description{
SRS of 150 members of the Statistical Society of Canada, downloaded from ssc.ca in August, 2006.
}
\references{
Lohr (2021), Sampling: Design and Analysis, 3rd Edition. Boca Raton, FL: CRC Press.

Lu and Lohr (2021), R Companion for \emph{Sampling: Design and Analysis, 3rd Edition}, 1st Edition. Boca Raton, FL: CRC Press.
}
\keyword{datasets}
