% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teachmi.R
\docType{data}
\name{teachmi}
\alias{teachmi}
\title{teachmi data}
\format{
This data frame contains the following columns:

\describe{
\item{dist:}{school district size: large or sm/me}

\item{school:}{school identifier}

\item{popteach:}{number of teachers in that school}

\item{ssteach:}{number of surveys returned from that school}
 
 }
}
\usage{
data(teachmi)
}
\description{
Cluster sizes for data \emph{teachers}.
}
\references{
Lohr (2021), Sampling: Design and Analysis, 3rd Edition. Boca Raton, FL: CRC Press.

Lu and Lohr (2021), R Companion for \emph{Sampling: Design and Analysis, 3rd Edition}, 1st Edition. Boca Raton, FL: CRC Press.
}
\keyword{datasets}
