% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teachnr.R
\docType{data}
\name{teachnr}
\alias{teachnr}
\title{teachnr data}
\format{
This data frame contains the following columns:

\describe{
\item{hrwork:}{number of hours required to work at school per week}

\item{size:}{class size}

\item{preprmin:}{minutes spent per week in school on preparation}

\item{assist:}{minutes per week that a teacher's aide works with the teacher in the classroom}
}
}
\usage{
data(teachnr)
}
\description{
Data from a follow-up study of nonrespondents from Gnap (1995).
}
\references{
Gnap, R. (1995). Teacher Load in Arizona Elementary School Districts in Maricopa County.
\emph{Ph.D. dissertation.} Tempe, AZ: Arizona State University.

Lohr (2021), Sampling: Design and Analysis, 3rd Edition. Boca Raton, FL: CRC Press.

Lu and Lohr (2021), R Companion for \emph{Sampling: Design and Analysis, 3rd Edition}, 1st Edition. Boca Raton, FL: CRC Press.
}
\keyword{datasets}
