% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uneqvar.R
\docType{data}
\name{uneqvar}
\alias{uneqvar}
\title{uneqvar data}
\format{
This data frame contains the following columns:

x: x value
 
y: y value
}
\usage{
data(uneqvar)
}
\description{
Artificial data used in exercises of Chapter 11.
}
\references{
Lohr (2021), Sampling: Design and Analysis, 3rd Edition. Boca Raton, FL: CRC Press.

Lu and Lohr (2021), R Companion for \emph{Sampling: Design and Analysis, 3rd Edition}, 1st Edition. Boca Raton, FL: CRC Press.
}
\keyword{datasets}
