% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algebra.R
\docType{data}
\name{algebra}
\alias{algebra}
\title{algebra data}
\format{
This data frame contains the following columns:

\describe{
\item{class:}{class number}

\item{Mi:}{number of students \eqn{M_i} in class}

\item{score:}{score of student on test}
}
}
\usage{
data(algebra)
}
\description{
Fictional data for an SRS of 12 algebra classes in a city, from a population of 187 classes.
}
\references{
Lohr (2021), Sampling: Design and Analysis, 3rd Edition. Boca Raton, FL: CRC Press.

Lu and Lohr (2021), R Companion for \emph{Sampling: Design and Analysis, 3rd Edition}, 1st Edition. Boca Raton, FL: CRC Press.
}
\keyword{datasets}
