% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anthsrs.R
\docType{data}
\name{anthsrs}
\alias{anthsrs}
\title{anthsrs data}
\format{
This data frame contains the following columns:

finger:length of left middle finger (cm)
  
height:height (inches)
 
wt: sampling weight
}
\usage{
data(anthsrs)
}
\description{
Length of left middle finger and height for a SRS of size 200 from anthrop data
}
\references{
Lohr (2021). Sampling: Design and Analysis, 3rd Edition.
}
\keyword{datasets}
