% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{data}
\name{classes}
\alias{classes}
\title{classes data}
\format{
This data frame contains the following columns:

\describe{
\item{class:}{class ID number}

\item{class_size:}{number of students in class}
}
}
\usage{
data(classes)
}
\description{
Population sizes for 15 classes, used in Chapter 6 of SDA to illustrate
unequal-probability sampling.
}
\references{
Lohr (2021), Sampling: Design and Analysis, 3rd Edition. Boca Raton, FL: CRC Press.

Lu and Lohr (2021), R Companion for \emph{Sampling: Design and Analysis, 3rd Edition}, 1st Edition. Boca Raton, FL: CRC Press.
}
\keyword{datasets}
