% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classpps.R
\docType{data}
\name{classpps}
\alias{classpps}
\title{classpps data}
\format{
This data frame contains the following columns:

class: Class ID number

class_size: Number of students in class

finalweight: Sampling weight for student

hours: Number of hours spent studying statistics
}
\usage{
data(classpps)
}
\description{
Two-stage unequal-probability sample without replacement from the population of classes in classes data
}
\references{
Lohr (2021). Sampling: Design and Analysis, 3rd Edition.
}
\keyword{datasets}
