% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coots.R
\docType{data}
\name{coots}
\alias{coots}
\title{coots data}
\format{
This data frame contains the following columns:

clutch: Clutch number from which eggs were subsampled.

csize: Number of eggs in clutch ( M i )

length: length of egg (mm)

breadth: maximum breadth of egg (mm)

volume: calculated as 0.000507*length * breadth

tmt: = 1 if received supplemental feeding, 0 otherwise
}
\usage{
data(coots)
}
\description{
Selected information on egg size, from a larger study by Arnold (1991). Data
provided courtesy of Todd Arnold. Not all observations are used for this data set, so results
may not agree with those in Arnold (1991).
}
\references{
Lohr (2021). Sampling: Design and Analysis, 3rd Edition.
}
\keyword{datasets}
