% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deadtrees.R
\docType{data}
\name{deadtrees}
\alias{deadtrees}
\title{deadtrees data}
\format{
This data frame contains the following columns:

\describe{
\item{photo:}{number of dead trees in plot from photograph}

\item{field:}{number of dead trees in plot from field observation}
}
}
\usage{
data(deadtrees)
}
\description{
Number of dead trees recorded by photograph and field count for a (fictional) SRS of 25 plots taken from a population of 100 plots.
}
\references{
Lohr (2021), Sampling: Design and Analysis, 3rd Edition. Boca Raton, FL: CRC Press.

Lu and Lohr (2021), R Companion for \emph{Sampling: Design and Analysis, 3rd Edition}, 1st Edition. Boca Raton, FL: CRC Press.
}
\keyword{datasets}
