% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpa.R
\docType{data}
\name{gpa}
\alias{gpa}
\title{gpa data}
\format{
This data frame contains the following columns:
\describe{
\item{suite:}{suite (psu) identifier}

\item{gpa:}{grade point average of person in suite}

\item{wt:}{sampling weight, = 20 for every observation}
}
}
\usage{
data(gpa)
}
\description{
GPA data from Chapter 5 of SDA.
}
\references{
Lohr (2021), Sampling: Design and Analysis, 3rd Edition. Boca Raton, FL: CRC Press.

Lu and Lohr (2021), R Companion for \emph{Sampling: Design and Analysis, 3rd Edition}, 1st Edition. Boca Raton, FL: CRC Press.
}
\keyword{datasets}
