% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htsrs.R
\docType{data}
\name{htsrs}
\alias{htsrs}
\title{htsrs data}
\format{
This data frame contains the following columns:

\describe{

\item{rn:}{random number used to select unit}

\item{height:}{height of person, cm}

\item{gender:}{
M = male

F = female}
}
}
\usage{
data(htsrs)
}
\description{
Height and gender for an SRS of 200 persons, taken from \emph{htpop} data
}
\references{
Lohr (2021), Sampling: Design and Analysis, 3rd Edition. Boca Raton, FL: CRC Press.

Lu and Lohr (2021), R Companion for \emph{Sampling: Design and Analysis, 3rd Edition}, 1st Edition. Boca Raton, FL: CRC Press.
}
\keyword{datasets}
