% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htstrat.R
\docType{data}
\name{htstrat}
\alias{htstrat}
\title{htstrat data}
\format{
The columns and names are as in htsrs data.
}
\usage{
data(htstrat)
}
\description{
Height and gender for a stratified random sample of 160 women and 40 men,
taken from htpop data.
}
\references{
Lohr (2021). Sampling: Design and Analysis, 3rd Edition.
}
\keyword{datasets}
