% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\docType{data}
\name{impute}
\alias{impute}
\title{impute data}
\format{
This data frame contains the following columns:

person: identification number for person

age: age in years

gender: M=male, F=female

crime: = 1 if victim of any crime, 0 otherwise

violcrime: = 1 if victim of violent crime, 0 otherwise
}
\usage{
data(impute)
}
\description{
Small artificial data set used to illustrate imputation methods.
}
\references{
Lohr (2021). Sampling: Design and Analysis, 3rd Edition.
}
\keyword{datasets}
