% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/santacruz.R
\docType{data}
\name{santacruz}
\alias{santacruz}
\title{santacruz data}
\format{
This data frame contains the following columns:

tree: Tree number

seed92: Number of seedlings in 1992

seed94: Number of seedlings in 1994
}
\usage{
data(santacruz)
}
\description{
The number of seedlings in the sampled psus on Santa Cruz Island, California, in 1992 and 1994.
}
\references{
Lohr (2021). Sampling: Design and Analysis, 3rd Edition.
}
\keyword{datasets}
